New Enemy Sprites, Fixed Player Sprite, HealthBar, Jumping Fix, Health Fix. Look at description for detail
Added Custom Floating Enemy Sprites and implimented them, fixed the player sprite so idle was the idle animation instead of a crouch animation, and so the sprites were cut out better. Made it so there is a working healthbar, and fixed the health script so it uses the health bar in start, instead of only using the health bar after we are damaged.